chore: convert booking_denormalized e2e test to integration test
